minor bugs fixed in counts type

+++ endpoints/00-domain/src/main/scala/ptpger/domain/Counts.scala
@@ -4,11 +4,11 @@ import io.circe.generic.JsonCodec
 
 @JsonCodec
 case class Counts (
-    count: Int,
-    `new`: Int,
-    inProgress: Int,
-    completed: Int,
-    onHold: Int,
-    rejected: Int,
-    approved: Int,
+    count: Long,
+    `new`: Long,
+    inProgress: Long,
+    completed: Long,
+    onHold: Long,
+    rejected: Long,
+    approved: Long,
 )
